added a lot more tests and changed goal_views to float64 so that it can unmarshal scientific notation data format coming from es. :=((

+++ delib/types.go
@@ -27,10 +27,10 @@ type Unit struct {
 	ChannelUrl         string     `json:"channel_url,omitempty"`
 	Status             string     `json:"status,omitempty"`
 	GoalPeriod         string     `json:"goal_period,omitempty"`
-	GoalViews          uint64     `json:"goal_views,omitempty"`
+	GoalViews          float64    `json:"goal_views,omitempty"`
 	Duration           uint32     `json:"duration,omitempty"`
-	Clicks             uint64     `json:"clicks,omitempty"`
-	Views              uint64     `json:"views,omitempty"`
+	Clicks             float64    `json:"clicks,omitempty"`
+	Views              float64    `json:"views,omitempty"`
 	ExcludedCategories []string   `json:"excluded_categories,omitempty"`
 	Devices            []string   `json:"devices,omitempty"`
 	Categories         []string   `json:"categories,omitempty"`
